fix: 월렛 계좌 차감 요청 타입과 충돌 예외 처리 수정

+++ coreBanking/src/main/java/woorifisa/project/coreBanking/domain/accountTransaction/dto/request/DebitWalletAccountRequest.java
@@ -4,6 +4,6 @@ public record DebitWalletAccountRequest(
         String walletChargeRequestId,
         Long customerId,
         Long withdrawAccountId,
-        Long chargeAmount
+        Integer chargeAmount
 ) {
 }
